color_class: Remove unused library.
This library was used to determine if a stream color was dark or
not and add `dark_background` class to it. With the new stream
background colors we don't need this anymore. The css for this
is still used private message which will be removed in a future
commit when we redesign PMs.